chore: use uint16 for output index
We were using uint8 for output index which limits the tx output to
around 256. For DEX batch settlement this seems to be small. This commit
extends the field to uint16 and cap the output index to limit of
TX_OFFSET (10000).

- issue: https://github.com/omisego/plasma-contracts/issues/168

+++ plasma_framework/contracts/src/utils/UtxoPosLib.sol
@@ -47,9 +47,9 @@ library UtxoPosLib {
     function outputIndex(UtxoPos memory _utxoPos)
         internal
         pure
-        returns (uint8)
+        returns (uint16)
     {
-        return uint8(_utxoPos.value % TX_OFFSET);
+        return uint16(_utxoPos.value % TX_OFFSET);
     }
 
     /**
